add pro banner to pitr settings
