+++ dashboard/src/features/orgs/projects/database/settings/components/UpgradeNotification/index.ts
@@ -1,0 +1,1 @@
+export { default as UpgradeNotification } from './UpgradeNotification';
